fixes the way we handle assay types in the add_genetic_id function
basically we must first filter to just the plate run at hand, then
we can make use of the `names_expand` argument to pivot so that we dont
have to do the hacky "DELETE_ME" row. Note that to do this we must first
make the assay_type into a factor to let `name_expand` know there are
other values not present here that need to be expanded.